Refactoring some names
