--- src/utils/jsonwrap.js
@@ -1,5 +1,0 @@
-import {jsonWrap as regex} from "./regex.js";
-
-export function jsonWrap (arg) {
-	return regex.test(arg);
-}
