+++ src/utils/regex.js
@@ -2,7 +2,6 @@ export const bodySplit = /&|=/;
 export const collection = /(.*)(\/.*)$/;
 export const hypermedia = /(([a-z]+(_)?)?id|url|uri)$/i;
 export const mimetype = /;.*/;
-export const jsonWrap = /^["[{].*["}\]]$/;
 export const trailing = /_.*$/;
 export const trailingS = /s$/;
 export const trailingSlash = /\/$/;
